Refactor WebGL shader initialization and uniform caching
- Extract and centralize shader compilation and program setup logic in `Renderer`.
- Add `onProgramCreated` hook for uniform caching across subclasses.
- Refactor uniform caching in `RosslerRenderer`, `MandelbrotRenderer`, `JuliaRenderer`, and `RiemannRenderer` to leverage base implementation.
- Improve code clarity by removing redundant `updateUniforms` calls and consolidating GL resource cleanup into `Renderer.destroy()`.
