+++ src/renderers/RosslerRenderer.js
@@ -56,20 +56,23 @@ export class RosslerRenderer extends FractalRenderer {
         return fragmentShaderRaw.replace('__MAX_ITER__', this.MAX_ITER).toString();
     }
 
+    /**
+     * Called after GL program is created.
+     * Caches Rossler-specific uniform locations.
+     * @override
+     */
     onProgramCreated() {
-        // No additional resources needed (no orbit texture like Mandelbrot)
-    }
+        super.onProgramCreated();
 
-    needsRebase() {
-        return false; // No perturbation orbit
-    }
-
-    updateUniforms() {
-        super.updateUniforms();
+        // Rossler-specific uniform locations
         this.paramsLoc = this.gl.getUniformLocation(this.program, 'u_params');
         this.frequencyLoc = this.gl.getUniformLocation(this.program, 'u_frequency');
         this.phaseLoc = this.gl.getUniformLocation(this.program, 'u_phase');
         this.iterationsLoc = this.gl.getUniformLocation(this.program, 'u_iterations');
+    }
+
+    needsRebase() {
+        return false; // No perturbation orbit
     }
 
     draw() {
